pyproject.toml
[build-system]
requires = [ "setuptools",]
build-backend = "setuptools.build_meta"

[project]
name = "ckanext-composite-search"
version = "0.4.0"
description = ""
classifiers = [ "Development Status :: 4 - Beta", "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)", "Programming Language :: Python :: 3",]
keywords = [ "CKAN",]
dependencies = []
authors = [
    {name = "DataShades", email = "datashades@linkdigital.com.au"},
    {name = "Sergey Motornyuk", email = "sergey.motornyuk@linkdigital.com.au"},
]
maintainers = [
    {name = "DataShades", email = "datashades@linkdigital.com.au"},
]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.license]
text = "AGPL"

[project.urls]
Homepage = "https://github.com/DataShades/ckanext-composite-search"

[project.optional-dependencies]
test = [ "pytest-ckan",]

[project.entry-points."ckan.plugins"]
composite_search = "ckanext.composite_search.plugin.base:CompositeSearchPlugin"
default_composite_search = "ckanext.composite_search.plugin.default:DefaultSearchPlugin"

[project.entry-points."babel.extractors"]
ckan = "ckan.lib.extract:extract_ckan"

[tool.setuptools.packages]
find = {}

[tool.black]
line-length = 79
preview = true


[tool.pytest.ini_options]
filterwarnings = [ "ignore::sqlalchemy.exc.SADeprecationWarning", "ignore::sqlalchemy.exc.SAWarning", "ignore::DeprecationWarning",]
addopts = "--ckan-ini test.ini"
